+++ apps/vs-code-designer/src/app/utils/vsCodeConfig/tasks.ts
@@ -223,7 +223,7 @@ async function overwriteTasksJson(context: IActionContext, projectPath: string):
             options: {
               cwd: debugSubpath,
               env: {
-                PATH: '${config:azureLogicAppsStandard.dependenciesPath}\\\\NodeJs;${config:azureLogicAppsStandard.dependenciesPath}\\\\DotNetSDK;$env:PATH',
+                PATH: '${config:azureLogicAppsStandard.autoRuntimeDependenciesPath}\\\\NodeJs;${config:azureLogicAppsStandard.autoRuntimeDependenciesPath}\\\\DotNetSDK;$env:PATH',
               },
             },
             problemMatcher: '$func-watch',
@@ -256,7 +256,7 @@ async function overwriteTasksJson(context: IActionContext, projectPath: string):
             args: ['host', 'start'],
             options: {
               env: {
-                PATH: '${config:azureLogicAppsStandard.dependenciesPath}\\\\NodeJs;${config:azureLogicAppsStandard.dependenciesPath}\\\\DotNetSDK;$env:PATH',
+                PATH: '${config:azureLogicAppsStandard.autoRuntimeDependenciesPath}\\\\NodeJs;${config:azureLogicAppsStandard.autoRuntimeDependenciesPath}\\\\DotNetSDK;$env:PATH',
               },
             },
             problemMatcher: '$func-watch',